docs: move pages to src folder
