+++ src/pages/notes/[id].jsx
@@ -4,6 +4,10 @@ import { useRouter } from 'next/router'
 const Page = () => {
   const router = useRouter()
   const {id} = router.query
+  
+  // const {params} = router.query
+  // params === ['a','b','c']
+  // route notes/[...params].jsx
 
   return(
     <div>
