+++ src/pages/notes/index.js
